Factor main.rs apart into a few separate files
Having everything in main.rs is a bit unwieldy
